Allow rendering nodes directly, render nodes from EssenceNode
This refactors the wrapper and node partials for menus such that they
can be rendered using Rails' implicit object rendering mechanism.

If a node is rendered in a partial using `<%= render node %>`, it will
not render itself by default. However, if you pass `render_self: true`,
it will render itself, aka the root node:

```
<%= render node, options: { render_self: true } %>
```

This will generally be what you want when you use a node in an Essence;
however, often you'll want to render just the children. This is the
default that happens when you run:

```
<%= render node %>
```

The `render_menu` helper renders just the children by default, but now
will also accept `render_self`:

```
<%= render_menu('main_menu', render_self: true) %>
```