Add errors when node cant be deleted (#1828)
* Add error before destroying node with attached essence

Nodes that are used on e.g. content pages can not be simply destroyed.
This commit adds an error message so the user knows what's going on.

* Do not destroy nodes with dependent nodes that are in use on a page

This moves the validation up in the file so that awesome_nested_set does
not run its `delete_all` callback before we validate that it can do
that.

It also extends the validation such that we collect all the pages that
are referenced in the subtree below this node.