Remove usages of visible attribute from Page
This is a legacy attribute that was used to show the page in the navigation. Since the render_navigation helper has been removed and the feature has been replaced with a much better feature (Menus) we do not need this attribute anymore.

Also this attribute was used to generate the url of a page only taking visible parents into account. This also has been removed.