Install alchemy npm package into host app
Instead of compiling the pack in a hacky way we install the alchemy/admin npm package as dependency of the host apps package.json and copy the admin entry point into the host apps packs folder.

This reduces the amount of hacks we need in Alchemy in order to make webpack work.