Set proper nested set scope on page (#1837)
* Set proper nested set scope on page

* Ensure layoutpages factory never creates a parent

With the awesome nested set scope set to language and layoutpage there is no parent with the same scope to move the newly created page to.

* Ensure page factory sets language from parent

If the page has a parent set we use that as the language. We must avoid to call the parent method in the factory, that's why we use the @cached_attributes here.

* Fix page language_root factory for when language is nil

We might set language to nil in tests explicitly.

* Refactor page specs after factory changes

Now that the page factory sets the language to the parent if necessary we can rewrite some specs.