Backport ContentEditor to 4.6, deprecate removed methods on `Alchemy::Content` (#1847)
* Backport Content Editor from Master

In order to deprecate the removed methods from the content model in the
master branch, we need to introduce the content editor here.

* Re-add form_field_[name,id] to Alchemy::Content

These were removed in the master branch with the introduction of the ContentEditor, but gems
in the ecosystem still use them. In order to provide a good upgrade path
to 5.0, we should reintroduce and deprecate these methods.