Introduce page.url_path and use it for alchemyPageSelect (#1859)
* Introduce page url_path service class

Use this class to generate the url_path to a page.

It takes several circumstances into account

1. It returns just a slash for language root pages of the default langauge
2. It returns a url path with a leading slash for regular pages
3. It returns a url path with a leading slash and language code prefix for pages not having the default language
4. It returns a url path with a leading slash and the language code for language root pages of a non-default language

* Return the page url_path with API responses

We want to use the newly introduced url_path attribute on the page selector in the admin. It in general makes sense to have the correct url to a page from the API anyway.

* Remove @url_prefix

This variable is not used anymore. since we generate the url in page links via the API response.

* Use Page#url_path in alchemyPageSelect

We do not need to care about adding leading slashes in the template anymore and it fixes a lot of missing features for multi language pages.

* Use page.url_path in node.url

Instead of having (buggy) page url generating code in the node model, we use the newly introduced page.url_path attribute that takes all necessities into account.

* Appease Rubocop

* Use page.url_path in page link dialog

And with that fix bugs for wrongly generated urls in multi language environments.