Show url name in Page tree (#1856)
* Show url_path of page in admin page tree

The url_path of a page is useful information. We should always display it not only if the page redirects to external.

* Format sitemap.scss with Prettier

* Change the sitemap urlname background color

The old color was too prominent for that information.

* Show a legend on top of admin page trees

* Use a left aligned column layout for page tree

* Slighty more space for node page name and url

* Eager load associated records in page tree serializer

Now that we need the language and its site we should eager load them in the page tree serializer. Also the locker of a page should be eager loaded.

On a site with ~300 pages loading the page tree went from 722ms to 221ms on my machine.