Change default url format to basename+extension
The default of the url format in dragonfly is `:name` what is an alias to `:basename.:ext`. We used to pass `Picture#name` (without the suffix) into the ragonfly `url` method. Since we now extracted the url generation out of the picture class it might be possible to pass a dragonfly attachment (generated by PictureVariant) into the url class. Leading to duplicated extensions in the url, because dragonflys name is basename+ext