Create default language in host app's locale
Our default language factory will create a page in German. That's not
great for apps that do not have German as an `available_locale`, as
Alchemy now validates the language_code of the language. This change
adds a new trait "german" to satisfy our existing test suite, but
creates by default a language in the host apps first available locale.