Move back to Travis CI (#1907)
* Move back to Travis CI

This reverts commit 411fc9d548d46b76f6e135a85d2966984b875fce.

* Update codeclimate_test_reporter in travis.yml

* Remove warnings from travis.yml

* Update travis build matrix

This with all supported Ruby and Rails versions

* Set up travis jobs for rspec and jest specs

* Skip coverage upload for Jest specs

It is not trivial to merge the coverage with the RSpec coverage. Since we do not have much Jest specs yet, it is not worth it in the moment.

* Adjust rails generators coverage filter

We do not test generators and after we moved the generators out of the lib/rails folder we need to adjust the filter.