Add element definition api (based on PageLayout definitions) (#1986)
Adds a ElementDefinition repository class (like PageLayout) that allows elements to be registered outside of the elements.yml file.

This is intended to be used in modules or extensions so that elements offered don't have to be added to the elements.yml manually.

To ensure backwards compatability, the Element::Definitions API is largely unchanged and just delegates to ElementDefinition. ElementDefinition is an adjusted version of the concern with the general API of PageLayout on top.