Do not change page_id on reorder elements
Reordering elements does not need to change the elements page id anymore
since we removed the trash.

This time we do not change the order of the element ids passed via the params.
To verify I updated the test so that the elements get created upfront and
the order is explicitely assured.