Fix folder toggle for Menu
In the previous commits, NodeTree was supported with type
feature such that it could be used generically. This caused
an issue where existing Menu was not able to toggle.

This commit updates node_folder to include data-record-id
rather than data-node-id to fix the issue and align with the
recent changes.