Add namespace for Tree related routes (#2037)
NodeTree is being used for rendering the Menu in tree
structure. This Tree could be made generic so that it could be
used at other places too. All that will be required is the fields
similar to what we are using for Node, and the API endpoints.
The API endpoints needs to be similar to what we have for Node.
To differentiate and use the existing `NodeTree.js`, a namespace could
be added so that it could be extended and new routes could be added
it someone needs to make a similar Tree structure anywhere else too.

This commit adds a namespace for existing Alchemy.routes for node type
objects. The type is added as "node"  at `NodeTree.js`, `node_folder.hbs` and
`_node.html.erb` since these are the places from where the type data
could be sent out and the current NodeTree works as expected.
Similarly, in future if it is required to use the currently implemented
`NodeTree`, all that will be required is to extend and add the routes
to the API with namespace, update the model similar to the Node model
attributes and use `Alchemy.NodeTree()`.