Make record generic for picture assignment
In the previous commit, a basic foundation was added for using
generic picture assignment box. The trial was added for Book module
of Alchemy Bookstore.

This commit removes the book dependency and adds record so that
the assignment could be generic. To support this, the changes were
made at all the instances where admin_pictures/index action is getting
called. The record_type is also added so that the record could be
any of the models and can be used in retrieving the record from db.