Add picture_assignment_redirect_url helper for fetching assignment url
Before this commit, the assignment feature for Alchemy::Bookstore module
was added. This was made generic and changes were made at the overlay
files to test the working. This caused the essence picture assignment
to break since it is using the same route.

This commit adds a helper which fetches the urls based on the arguments
passed. If the content is present, the essence_picture assignment
path is returned or else the picture_assignment url.