Add render_picture_assignment_box helper to render the picture assignment box
The primary goal of Picture abstraction is to make the picture
assignment flexible for the users with the present image archive.
This commit adds a helper method called render_picture_assignment_box
which renders picture_assignment partial which comes with a number
of helpful dialog boxes facilitating the selection of images as
required. It also helps us upload the image and then mark the selection.