Create defined contents if they are missing
If new contents have been defined on the elements definition, but they are
not existing yet we create them on demand.

This happens during rendering (speak loading) elements in the elements window
during page edit. Ideally we would only instantiate them, but since the content
editors expect a persisted content we cannot do that.

This also closes #2031