Destroy public version if public checkbox is unset
In order unpublish a page via the "Published" checkbox
in the page settings dialog we need to destroy it if
the value transmitted is en empty string (or nil).

If we would delegate the date to the public version we would
make that version to the current draft version. That is not
what we want, so we destroy the public version and keep the
current draft version.