Scope Node Trees by Language and Menu Type
We want to have separate node trees for each menu in a language.

Unfortunately, this piece of configuration is extremely hard to test,
because `awesome_nested_set` is very robust when it comes to wrong
configurations.

I would have to set up a rather contrived scenario where I create two
root nodes, give children to each of them, assign incorrect menu_type
columns, and then use some ANS method to find out something that's
wrong. This is extremely unlikely given normal use of our classes, so I
prefer to think of this as just maintenance.