Remove visit action from admin pages controller
The benefit of this action was that it would unlock the page before
allowing the user to visit it. I think that's not necessarily a great
user experience, as an editor will probably only use the visit link to
quickly glance at the page without Alchemy present, and might not be
aware that this removes their lock on the page.