Add an Alchemy::EssenceHeadline
This adds an Alchemy::EssenceHeadline. It has the following features:
When used as a preview text, it will show the level of the headline so
that content editors can see what levels they are using. There are is
also an optional `level` element.

Headlines can be configured through their content settings. One can
restrict the sizes and levels available to an element, by specifying them
like follows:

```yml
- name: MyComponent
  contents:
     - name: my_headline
       type: EssenceHeadline
       settings:
          levels: [3, 4]
          sizes: [3]
```

If, as in the previous example, `sizes` only has one option, the select
input in the element editor will not be displayed.

If `sizes` is unconfigured, the default will be an empty array, and the
essence will have no size class.