Touch Elements only after update
When an essence has a related object, such as a picture or an
attachment, we need to break caches when that related object changes.
This reflects this desire - but stops touching the elements on first
creation. This results in a major speedup when publishing pages, as
publishing results in a lot of essence and element creation.