Add a fast element deletion service class
Deleting a collection of elements is rather slow at the moment. As
elements as well as contents have a `dependent: :destroy` on them, every
dependent record needs to be loaded and destroyed individually.

This class uses the knowledge we have of elements, contents and essences
in order to be able to destroy a collection of elements in a maximum of
less than 20 database calls (if the collection of elements uses a lot of
different essences, more calls happen, but there's only a limited amount
of essence classes).

There's one additional call to the database in here as a safeguard so we
do not produce orphaned elements. It's fast, and gives a sense of
security.