Start Element indexes with 1
This makes sure the element indexes start with 1, as `acts_as_list`
expects.

It also wraps the creation of nested elements in the `DuplicateElement`
class in an `Element.acts_as_list_no_update` block. The code in here is
safe, and when called from e.g. Alchemy::Element#copy, it makes it
faster, too.