Fix execjs to 2.7.0 for developers and CI (#2094)
There's an ecosystem bug in the new release of `execjs`, where the new
version (after five years of stability) breaks a few libraries. In our
case, it's `handlebars-assets`.

I'm opting not to include this fix in the .gemspec file, as I expect
this to be rectified pretty soon, and it's not our responsibility to
version control transient dependencies.