Pass crop parameter in default EssencePicture#picture_url_options (#2098)
Prior to this commit, calling `my_essence.picture_url` without
parameters would return an uncropped image be default. In order to get
the cropped picture, one would have to add `my_essence.picture_url(crop:
true)`. This is unintuitive: For an EssencePicture that I have cropped
as a user (crop values are present), `essence.picture_url` as well as
`essence.ingredient` should return the cropped version of the image.