Stick with Dragonfly 1.3 for now
With Dragonfly 1.4 we cannot use `image_file.convert`
anymore, since this raises an exception telling us to use
`Dragonfly::ImageMagick::Commands#convert` instead,
but this turns does not work as expected:

1. We will have multiple conversion per image for the same arguments
2. The thumbnails do not get cropped anymore

We only use `convert` because Dragonfly does not support cropping and resizing in one covert command like we do for performance and image quality reasons.

One possible solution would be to create our own processor that we then can make use of.

Let's stick to 1.3 for now.