Assign, crop and remove files and pictures client side (#2069)
* Add an admin picture url end point.

This returns the URL to processed image.
It accepts the same transformation options then the PictureUrl class

* Set crop values on essence in memory

Instead of updating the essence picture on crop apply
we just set the crop values on hidden fields that get
posted when the element gets safed.

* Extract image loader into ES6 module

* Use image loader in thumbnails handler

* Assign pictures and files in memory

Instead of assigning pictures and files to an essence via ajax
we just set the id on the form field that gets submitted with
the element on save.

* Move removePicture into thumbnail ES module

* Return default thumbnail options if no image assigned

We want to be able to render a thumbnail even if there is no image assigned yet.

* Remove flaky useless file overlay tag filter feature specs

These tests always have been really flaky and there is not much value
in wasting Carbon keeping them.

* Do not listen for disabled dialog links

If a `disabled` class is added to a link that has dialog options we ignore that.

* Toggle crop link from JS

Since we now toggle the image id and update the thumbnail
from JS we need to handle the crop link
active state as well.

* Load picture to crop from params

Instead of loading the image to crop from the essence we pass the id via params.

This allows to crop a picture without assigning it first.

* Update cropLink after changing image dynamically

We now always have the image when chosing an image and
click the crop link.

* Update Jcrop to v0.9.15

This resolves some loading issues.

* Add remove file link to essence_file editor

One should be able to unassign a file from the element.

* Try to reduce the cognitive load of pictureEditors

Let's try to make code climate happy.

* Smoother image loading after update

* Convert ImageCropper into ES6 class

* Use crop values from hidden fields instead of database

That way the cropper can be used several times w/o persisting the values first.

* Initialize file and picture editors after creating element

* Move image_cropper_settings into EssencePicture

These settings are better on the model instance,
instead of in the controller. That way we can read them even before we load
the controller action.

* Assign picture to essence if cropping

amend load picture from params?

* Reset image cropper field values to default box

* Enable crop link even if now crop values present

For a newly assigned picture we do not have the values yet.

They get populated during cropping anyway, so we do not need that check.

* Refactor image cropper

CodeClimate wants us to reduce complexity

* Extract image cropper settings into service class

This class is free from any picture, essence or content
dependencies and can be used by any model that can
provide necessary values for the image cropper.

* Set image file width and height on picture editor

We need this value in JS if we want to display the thumbnail cropped
with its default mask on initial render

* Move crop mask calculation into JS

* Move initial Box into cropper.js

* Remove TrashWindow refresh from dragndrop handler

The TrashWindow has been removed

* Pass form_field_id with search and filter params

We use these params in several places (tag list, filter drop down, full
text search) in the archive overlays. Otherwise we will loose the
value while searching/filtering the archive overlay.

* Only enable cropper if configured in the element settings

* Pass the desired render size to image thumbnail

Use the thumbnail_size only as fallback

* Fix image size discrepancy in article element

The setting and the view were using contrary settings

* Fit the image in picture editor thumbnail

Use object-fit: contain to make sure that the images in the essence picture
editor are always using the maximum of space avaible while maintaining the
aspect ratio.

* Add header and slider styles

Make it easier to review visual changes to the cropper in the dummy app

* Lower image render delay

* Do not send crop values if not enabled

* Make fileEditors work with essence_video and audio

* Display empty essence file as empty

* Fix image replacement spinner in Firefox

Firefox thinks the image is already complete if the src is removed.

Forcing the spinner in this case.

* Reset thumbnail alt tag on picture change

* Always use thumbnail_size for thumbnail sizes

* Fix initial thumbnail options

We need to set default crop from and default crop size right, and then
ImageMagick will do the right thing (tm).

* Move size_when_fitting into image_cropper_settings

* Remove transformations mixin from essence picture

We do not need any of these methids anymore.

* Simplify ImageCropperSettings

The image cropper settings do not need to do any calculations. They just
need to add the default crop from and crop size up to get a default mask
that works for it.

* Move specs from ImageCropper spec to EssencePicture Spec

* Further refactor default_crop_size

Make code climate happy

* Use substitution for image title translation

That way the whole strings format can be changed

* Add title to admin/pictures#url response

* Replace title after new thumbnail loads

* Set correct target size on essence picture editor

The target size is the size we want to render the final image in.

If the element does not have a size set, we take the image_size of the original image.

This makes sure we can render the correct thumbnail

* Reset the crop values on image change

A new image needs to get the default box, otherwise the crop box is calculated
from old image sizes.

* Maintain aspect ratio if one dimension is missing

* Fix box calculation in image cropper

* Send default crop values if value empty

Co-authored-by: Martin Meyerhoff <mamhoff@gmail.com>