Only pass crop values if cropping is enabled
Before we enabled cropping even if cropping was disabled
only because we had crop values in the database. This is
confusing, inconsistent and error prone. Imagine an
essence picture having crop values, but the image cropper
is disabled because the element settings has been changed.
The image was still be cropped in those cases.

Now cropping has to be explicitely enabled in the element
settings to crop an image, regardless of the database state.