Ensure resource table ends before filter/tag bar
A wide resource table in combination with the sidebar for filter/tags made the table slip under the sidebar. As a consequence it would hide the edit/delete buttons and even some attributes.

This commit enforces the width of the resource-table-wrapper can always only be as wide as it doesn't slip under the library_sidebar.