Do not make Ingredient an abstract class
This causes a lot of issues during creating an ingredient
record, because Rails holds a lot of precautions for abstract
class. Since this is not necessary for STI to work and we
already have validations of all necessary attributes this
can safely be removed allowing us to simplify the process
of creating an ingredient record.