Create ingredient on element collection
In the element editor with have a feature that creates ingredients
that are defined but not created yet. Before we removed the abstract
class from Ingredient we needed to create the ingredient on the class
instead of the collection. Since the collection is eager loaded it has
to be updated with the newly created ingredient in order to prevent
the ingredient to be created twice (leading to unique index violation
errors).