Ensure I18n can find the translation key for boolean ingredients
Before this change we were passing a boolean type in to `I18n.t` which failed to find the according translation. This bubbled up to even raise an error when rendering the `admin/elements/_header` view partial when the boolean ingredient is used for the preview text.

Solution: Passing a `String` instead of `Boolean` to `I18n.t` fixes the issue.

Fixes #2175