Do not validate during create
If an element has ingredient validations
careating it first and calling valid? after that triggers the on update
validation callback. Resulting in returning the new template instead.

Solved by only building the element and saving it later. This does not trigger the on update callback.
To spare an extra save we set the elements position instead of moving it down.
The only unnecessary ave we could not prevent is if we paste from clipboard.
But this is acceptable.