don't use existing element names when migrating from cells
The upgrader writes new fixed elements to the elements.yml for each
cell definition it finds in the cells.yml file. If there is an element
with the same name as the cell, we get two elements with the same name:
the original one and the new fixed one. This leads to problems when you
are searching for specific elements by name in your code.

Now the upgrader checks if the name already  exists before we add a
new fixed element and add `_elements` if we find an existing elements
name.

Most of the time the cells with the same name as existing elements
are used as a vehicle to group multiple elements to get a more
complex element.