make the fixed element upgrader more robust
The cell name migrator does a good job when running multiple times in
the same process, because the class is already in memory and the cache
is working. Running it multiple times in different processes led to
errors. This new name finder always takes the current version of the
elements.yml and also checks if a fixed element with the correct cell
name already exists. This makes the upgrader idem potent again.