Add custom gutentag tag validations class
We support having tags with different casing to exist
side-by-side because of historical reasons (we used acts-as-taggable-on
back in the days). We already have a custom Gutentag.normaliser,
but never a custom validation. That leads to "already taken" errors
trying to save an element or page with tags having different casing.