Explicitly validate uniqueness without case sensitivity
In Rails 6.1, the default behavior of the uniqueness validator changed
from case-sensitive to case-insensitive. The Rails folk have made a
beautiful deprecation warning for this behavior, which we now see in all
our Rails 6 builds.

This PR explicitly specifies that we don't want case sensitivity on
URLs, host names, and language codes. This is in line with browser
behavior and implicit behavior on newer Rails versions.