Allow passing a different partial to `render_element`
This is useful when rendering elements on a different page than the one
they are on, for example when rendering something like a blog overview.

What we're doing here is making all the options for `render` explicit.
The `object` key will do the magic for the partial name as a local
variable, and the locals are now more explicitly passed.