Ensure all ingredients get created during migration
It might be the case that ingredients are defined on an element
that does not have a equivalent content.

We want to make sure that the element is exaclty is how it is defined in the elements.yml after the migration.

If the content could not be found, we just skip copying the data.