Add spec for ingredients migrator; fix preloading
This adds a spec for the ingredients migrator that tests its
functionality. It also serves to reproduce an error under Rails 6.1+,
where the ingredients migrator would fail at preloading essences with
the following error:

ActiveRecord::EagerLoadPolymorphicError: Cannot eagerly load the
polymorphic association.

Changing the `includes` to `preload` fixes that error.