Create new pages as first child of parent page
When creating a new page, the user will click on the `+` button next to
a parent page. Consider the following situation: The parent page already
has a lot of child pages. When the user returns to the page tree, they
will not find their newly created page unless they scroll alll the way
to the bottom...

What this does is it moves the page up to near where the user has their
attention.