Create draft version for public pages in page seeder
If you pass a public_on date to the page seeder,
Alchemy will create a public version (because of the
public_on= attribute writer). The before_create filter
kicks in too late, leading to public pages without a
draft version.

Tried to use different before filters in the page model,
but nothing worked. So I fixed the problem where it
happens in the page seeder.