Connect to writing database during thumbnail generation
In a multi db setup where you have a reading and a writing database
we need to make sure that we are using the writing database.

This is necessary, because we are writing (caching) picture
thumbnails in a GET request. Usually Rails would handle this for us
on POST, PATCH/PUT and DELETE requests.

ActiveStorage does the same thing to support multi database
setups.

Closes #2428