Use ISO Date format in flatpickr
Currently the date/time being displayed in flatpickr is (correctly) set
to the user's current timezone (as detailed in
https://github.com/flatpickr/flatpickr/issues/1532)

Where this breaks is when it's sent back to the server to update the
page. Flatpickr sends back the date in it's default format, which does
not include any timezone information. Rails then takes that timestamp
and assumes it's in UTC when saving it to the database.

Which means a user in UTC+10 can save a page to go public at 1PM local
time, and after saving, it will actually be set to 11PM local time
(== 1PM UTC)

By forcing the underlying string value in the form to adopt the ISO
standard, Rails can parse both the time AND timezone from the incoming
data, and ensure the timestamp is saved to the correct UTC time in the
database, fixing the bug