Render nested elements of collapsed element editors
We need to render nested elements even if the element is folded,
because we need the element present in the DOM for the feature
"click element in the preview => load and expand element editor".

Together with the previous commit where introduced collapsing
of all nested elements if we fold an element, the performance
is much better now.

This reverts commit 0d7a7f9bbab9a4d0b378f7a70e6a8b4ee63a4795