Define #preload_relations on Ingredients
In order to be able to preload ingredients' related object graphs, we
need to be able to tell the rails preloader how those graphs look.

This adds a method that can be used on for example an element like this:

```
def preload_ingredient_relations(element)
  element.ingredients.group_by(&:preload_relations).each do |preload_relations, ingredients|
    preload(records: ingredients.map(&:related_object).compact, associations: preload_relations)
  end
  element
end

def preload(records:, associations:)
  if Rails::VERSION::MAJOR >= 7
    ActiveRecord::Associations::Preloader.new(records: records, associations: associations).call
  else
    ActiveRecord::Associations::Preloader.new.preload(records, associations)
  end
end
```

Most notorious for n+1 problems is the picture ingredient that tends to
have a lot of thumbs that we actually need.