Fix DOM ids and labels of ingredient editors
The boolean ingredient editor label tag must
not have a `for` attribute since it nests the input
tag. That input does not need an `id` as it is
nested. Solves an issue with non-clickable
label on the checkbox.

The image editor label must not have a `for`
attribute as well, as there is no focusable form
element.