Picture Ingredient: Fix NaN error with free height
When indicating that a picture can be cropped, but needs to have a
certain height or width but leaves the other dimension free, current
Alchemy throws a `NaN` error because of float division by Zero.

This commit handles this case by returning nil if the calculated mask is
Not a Number.