Add PageMutex model
The idea is to provide a possibility to lock a page inside a transaction to prevent, that two (or more) processes working on the same page. This will be included into the Page::Publisher to prevent, that in parallel running worker are updating the same page and creating duplicate entries. This implementation is heavily inspired by Solidus Spree::OrderMutex.