Fix actioncable autoload problem
The CI test pipeline breaks for Rails 7.1. The eager loading results into a Zeitwerk error in the turbo-rails gem. The current solution is to include actioncable if the application is running in a Github Action until this bug is fixed.

Ref: https://github.com/hotwired/turbo-rails/issues/512
(cherry picked from commit 8fe1bb3326378c1a0d571eea69496b68dab4fec7)