Add alchemy-button web component
This allows to remove all the Alchemy.Buttons state management
code we have right now and let the browser do it for us.

This DOES not work in Safari as they do not support extending
existing HTML elements, but this is ok in this case. Maybe we will
provide a polyfill someday or just do not support this feature in Safari.

In order to update forms not using the `alchemy_form_form` helper,
replace all usages of `data-alchemy-button` with `is="alchemy-button"`.