Add upload - components
Add a combination of file upload components. The Uploader component is handling the file upload to the server. The Progress - component will show the overall upload progress including file sizes and the FileUpload will handling each file upload and is informing the Progress - component.

Also a Jest - helper is introduced, which is preparing the Alchemy config and provide global function to suppress console errors and mock XHR requests.