Fix tinymce component specs
We need the matchMedia mock, because - once again -
jsdom has not implemented it (*sigh). Otherwise the spec
will not even load, because Tinymce 5 needs this standard
HTML API do be defined.

The intersection observer spec does not work anymore.
The feature itself works in the browser, but not in Jest specs.
Skipping for now.