Build our own shoelace bundle with rollup
It is not possible to pin and download shoelace
with importmap-rails. Since it is more efficient to have
a bundle anyway and depency management via yarn
much easier than with importmap-rails we install it
via yarn and build it with rollup. This can easyliy be
automated with github actions, so we have best of
all worlds.