Clear unnecessary legacy URLs
When renaming a page back to a URL it used to have, we get a legacy URL
that is the same as the current URL of the page, which can lead to
infinite redirects. This deletes any legacy URLs that are the same as
the current URL name in a callback.